Прикрутил React-router к <App />. Перенес окно с диалогом в отдельный компонент
