--- src/App.css
@@ -1,5 +1,0 @@
-:local(.react_container) {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-}
