+++ src/components/containers/Dialog/styles.css
@@ -1,0 +1,5 @@
+:local(.react_container) {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+}
